Melhora a robustez e segurança da API
- **Precisão Financeira com BigDecimal:**
  Substituição de `Double` por `BigDecimal` em todas as entidades, DTOs
  e serviços que manipulam valores monetários. Isso garante a precisão
  em cálculos financeiros e evita erros de arredondamento.

- **Validação de Entrada (DTO Validation):**
  Implementação de validação nos DTOs de requisição (@Valid e Jakarta
  Bean Validation). Isso garante a integridade dos dados na entrada da API
  e fornece feedback claro ao cliente em caso de dados inválidos (400 Bad Request).

- **Controle de Concorrência (Lock Otimista):**
  Adição do controle de concorrência com `@Version` na entidade `Conta` para
  prevenir condições de corrida (race conditions) e inconsistência de
  dados em operações simultâneas (ex: saques ao mesmo tempo).

- **Tratamento de Exceções Centralizado:**
  Criação e aprimoramento do `GlobalExceptionHandler` para capturar e
  formatar respostas de erro de forma consistente, incluindo handlers para
  erros de validação (400) e conflitos de concorrência (409).

+++ src/main/java/com/senai/conta_bancaria_spring/application/dto/ContaResponseDTO.java
@@ -7,17 +7,19 @@ import com.senai.conta_bancaria_spring.domain.entity.ContaPoupanca;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ContaResponseDTO {
     private Long numero;
-    private Double saldo;
+    private BigDecimal saldo;
     private String tipoConta;
 
     // Campos específicos que podem ser nulos
     private Long limite;
-    private Double taxa;
-    private Double rendimento;
+    private BigDecimal taxa;
+    private BigDecimal rendimento;
 
     // Método de fábrica para converter a entidade em DTO
     public static ContaResponseDTO fromEntity(Conta conta) {
